Major refactoring to reduce code duplication and simplify it
Derivatives no longer treat variables as parameters on the curve, and
are instead partial derivatives.
Uses the stl functional primitve operations and custom transcendental
operations to reduce code duplication of the eval and subs
implementations for the unary and binary operations.
Change expr_domain and binary_expr_domain to be aliases rather than
structs.
Implement template parameter pack version of eval - need to implement
it for subs as well